Update nbconvert_library.ipynb
When I try to use nbconvert as library, by following the documentation, the generated html file was always from 'lab' template. But when I passed the template_name as a keyword parameter to HTMLExporter, it worked and generated the html file with 'classic' template. So I changed the documentaion accordingly. Please do a review and accept the changes